Status specific links to Twitter were still broken, fixed.
The links behind "42 minutes ago" still showed some Twitpocalypse
left-overs.

svn path=/trunk/extragear/network/choqok/; revision=986680

+++ src/account.cpp
@@ -108,7 +108,7 @@ QString Account::apiPath() const
     return mApiPath;
 }
 
-QString Account::statusUrl( int statusId, const QString &userScreenName ) const
+QString Account::statusUrl( qulonglong statusId, const QString &userScreenName ) const
 {
     if( mServiceType == Twitter ) {
         return mStatusUrlBase.arg( QString::number(statusId), userScreenName );
